% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_biodata.R
\name{bind_biodata}
\alias{bind_biodata}
\title{Bind and Harmonize Biodiversity Data}
\usage{
bind_biodata(data_list)
}
\arguments{
\item{data_list}{List of dataframes (e.g., from GBIF and SPLink)}
}
\value{
A unified and standardized tibble.
}
\description{
Bind and Harmonize Biodiversity Data
}
