% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_label.R
\name{dedup_label}
\alias{dedup_label}
\title{Deduplicate Biodiversity Data}
\usage{
dedup_label(df)
}
\arguments{
\item{df}{Combined dataframe from bind_biodata.}
}
\value{
A dataframe with a new 'status_dedup' column.
}
\description{
Identifies duplicates based on collector, date, and coordinates.
It is flexible regarding column names to avoid errors.
}
