% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbif.R
\name{get_gbif}
\alias{get_gbif}
\title{Download occurrence data from GBIF}
\usage{
get_gbif(species, limit = 1e+05)
}
\arguments{
\item{species}{Character vector with scientific names.}

\item{limit}{Maximum records per species (default = 100000).}
}
\value{
Tibble of occurrence records standardized to 16 columns.
}
\description{
Download occurrence data from GBIF
}
