% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_splink.R
\name{get_splink}
\alias{get_splink}
\title{Download occurrence data from speciesLink}
\usage{
get_splink(species, api_key = Sys.getenv("SPLINK_API_KEY"))
}
\arguments{
\item{species}{Character vector of species names}

\item{api_key}{speciesLink API key (default: Sys.getenv("SPLINK_API_KEY"))}
}
\value{
Tibble of occurrence records standardized to 16 columns.
}
\description{
Download occurrence data from speciesLink
}
