% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_workflow.R
\name{run_workflow}
\alias{run_workflow}
\title{Full biodiversity workflow: GBIF + speciesLink + deduplication}
\usage{
run_workflow(species, export_path = getwd())
}
\arguments{
\item{species}{Character vector of scientific names.}

\item{export_path}{Character. Directory to save outputs.}
}
\value{
A list with two tibbles: [[1]] Labeled data, [[2]] Clean data.
}
\description{
Full biodiversity workflow: GBIF + speciesLink + deduplication
}
