% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_biodata.R
\name{bind_biodata}
\alias{bind_biodata}
\title{Bind GBIF and speciesLink data}
\usage{
bind_biodata(gbif_df, splink_df)
}
\arguments{
\item{gbif_df}{A tibble/dataframe returned by get_gbif().}

\item{splink_df}{A tibble/dataframe returned by get_splink().}
}
\value{
A single integrated tibble with 16 standardized columns.
}
\description{
This function combines standardized dataframes from GBIF and speciesLink
into a single integrated dataset.
}
