% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_label.R
\name{dedup_label}
\alias{dedup_label}
\title{Label duplicate biodiversity occurrence data}
\usage{
dedup_label(df)
}
\arguments{
\item{df}{Tibble with occurrence data}
}
\value{
Tibble with added DUP_GROUP and DUP_CRITERIA columns.
}
\description{
Label duplicate biodiversity occurrence data
}
