% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill_dedup.R
\name{kill_dedup}
\alias{kill_dedup}
\title{Remove marked duplicates}
\usage{
kill_dedup(df)
}
\arguments{
\item{df}{Dataframe returned by dedup_label.}
}
\value{
A cleaned tibble containing only unique records.
}
\description{
This function filters the dataframe to keep only 'Original' records
identified by dedup_label and removes the status column.
}
